+++ modules/Scheduling/includes/MasterScheduleReport.php
@@ -1,4 +1,12 @@
 <?php
+/**
+ * Master Schedule Report
+ *
+ * Included in ScheduleReport.php
+ *
+ * @package RosarioSIS
+ * @subpackage Scheduling
+ */
 
 //FJ multiple school periods for a course period
 /*$sections_RET = DBGet( "SELECT cs.TITLE as SUBJECT_TITLE,c.TITLE AS COURSE,cp.COURSE_ID,cp.PERIOD_ID,cp.TEACHER_ID,cp.ROOM,cp.TOTAL_SEATS AS SEATS,cp.MARKING_PERIOD_ID FROM COURSE_PERIODS cp,COURSES c,COURSE_SUBJECTS cs WHERE cs.SUBJECT_ID=c.SUBJECT_ID AND cp.COURSE_ID=c.COURSE_ID AND cp.SYEAR='".UserSyear()."' AND cp.SCHOOL_ID='".UserSchool()."'",array('PERIOD_ID' => 'GetPeriod','TEACHER_ID' => 'GetTeacher','MARKING_PERIOD_ID' => '_makeMP'),array('COURSE'));*/
@@ -29,8 +37,6 @@ $columns = array(
 	'STUDENTS' => _( 'Students' ),
 	'MARKING_PERIOD_ID' => _( 'Marking Period' ),
 );
-
-DrawHeader( ProgramTitle() );
 
 ListOutput( $sections_RET, $columns, 'Course Period', 'Course Periods' );